Code refactor and modularization
