+++ src/frontiers/frontier.py
@@ -1,0 +1,23 @@
+from abc import ABCMeta, abstractmethod
+
+from src.domain.state import State
+
+
+class Frontier(metaclass=ABCMeta):
+    @abstractmethod
+    def add(self, state: 'State'): raise NotImplementedError
+
+    @abstractmethod
+    def pop(self) -> 'State': raise NotImplementedError
+
+    @abstractmethod
+    def is_empty(self) -> 'bool': raise NotImplementedError
+
+    @abstractmethod
+    def size(self) -> 'int': raise NotImplementedError
+
+    @abstractmethod
+    def contains(self, state: 'State') -> 'bool': raise NotImplementedError
+
+    @abstractmethod
+    def get_name(self): raise NotImplementedError
