+++ src/searches/graphsearch.py
@@ -3,12 +3,15 @@ import os
 import os.path
 import sys
 import time
-import memory
 
-from action import Action
-from frontier import Frontier, FrontierIW
-from state import State
-from info import Info
+from src.frontiers.frontier import Frontier
+from src.frontiers.iw import FrontierIW
+
+from src.domain.state import State
+from src.domain.action import Action
+
+from src.utils import memory
+from src.utils.info import Info
 
 start_time = time.perf_counter()
 saved_once = False
@@ -18,19 +21,12 @@ def search(initial_state: State, frontier: Frontier):
     if not saved_once:
         save_run_information(None, None, None, {'Passed': False})
 
+    iterations = 0
     frontier.add(initial_state)
     explored = set()
 
     while True:
-
-        # iterations += 1
-        # if iterations % 1000 == 0:
-        #     print_search_status(explored, frontier)
-
-        # if memory.get_usage() > memory.max_usage:
-        #     print_search_status(explored, frontier)
-        #     print('Maximum memory usage exceeded.', file=sys.stderr, flush=True)
-        #     return None
+        # log_search_status(iterations, explored, frontier)
 
         if frontier.is_empty():
             if isinstance(frontier, FrontierIW):
@@ -65,6 +61,17 @@ def search(initial_state: State, frontier: Frontier):
                 frontier.add(expanded_state)
 
 
+def log_search_status(iterations, explored: set[State], frontier):
+    iterations += 1
+    if iterations % 1000 == 0:
+        print_search_status(explored, frontier)
+
+    if memory.get_usage() > memory.max_usage:
+        print_search_status(explored, frontier)
+        print('Maximum memory usage exceeded.', file=sys.stderr, flush=True)
+        return None
+
+
 def print_search_status(explored: set[State], frontier):
     status_template = ('#Expanded: {:8,}, '
                        '#Frontier: {:8,}, '
@@ -77,15 +84,35 @@ def print_search_status(explored: set[State], frontier):
 
 
 def save_run_information(explored: set[State], frontier, plan: list[list[Action]], information=None):
+    """
+    This function saves the information about the current run of the search algorithm.
+
+    :param explored: A set of states that have been explored by the search algorithm.
+    :param frontier: The frontier used by the search algorithm. It should have a size() method.
+    :param plan: A list of lists of actions representing the plan found by the search algorithm.
+    :param information: An optional dictionary containing additional information to be saved. If not provided, a default dictionary is created with the following keys:
+        - 'Passed': Always set to True.
+        - 'Expanded': The number of states expanded by the search algorithm.
+        - 'Frontier': The size of the frontier at the end of the search.
+        - 'Generated': The total number of states generated by the search algorithm.
+        - 'Solution length': The length of the plan found by the search algorithm.
+        - 'Time': The time taken by the search algorithm.
+    """
     elapsed_time = time.perf_counter() - start_time
     folder_path = Info.test_folder
 
+    # Check if the folder for saving the run information exists. If not, create it.
     if not os.path.exists(folder_path):
         os.makedirs(folder_path)
         print(f"Folder '{folder_path}' created.")
 
+    # The file path for saving the run information.
     file_path = os.path.join(folder_path, Info.test_name + '.json')
+
+    # Deserialize the existing data from the file.
     all_data = deserialize_from_json_file(file_path)
+
+    # If no additional information is provided, create a default dictionary.
     if information is None:
         information = {
             'Passed': True,
@@ -95,8 +122,11 @@ def save_run_information(explored: set[State], frontier, plan: list[list[Action]
             'Solution length': len(plan),
             'Time': elapsed_time
         }
+
+    # Add the information about the current run to the existing data.
     all_data[Info.level_name] = information
 
+    # Serialize the updated data back to the file.
     serialize_to_json_file(all_data, file_path)
